requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
flask>=3.0.0
flask-cors>=4.0.0
markdown>=3.4.0
weasyprint>=60.0
reportlab>=4.0.0

# RAG系统依赖
qdrant-client>=1.7.0
pymilvus>=2.3.0
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.24.0

